+++ order-payment/src/main/java/com/codecool/order_payment/dto/OrderItemDTO.java
@@ -3,5 +3,5 @@ package com.codecool.order_payment.dto;
 import java.math.BigDecimal;
 import java.util.UUID;
 
-public record OrderItemDTO(UUID itemPID, int amount, BigDecimal price) {
+public record OrderItemDTO(UUID itemPID, Long amount, BigDecimal price) {
 }
